zynqmp-starter: Add support for the KV260 AI vision starter kit
Add the first Xilinx board, along with update instructions.  This board
has no eMMC and all the secure boot variables are stored in the
ESP partition.

Note here that is using the downstream vendor version for u-boot and
TF-A.  This is an exception as we usually require upstream versions
for those projects.  However the xilinx patches are on the mailing list
for both of the projects waiting to be merged.  Once that happens we will
need to adjust our repos

Signed-off-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>