optee-os: set CFG_CORE_HEAP_SIZE=131072
On qemuarm64, the default TEE core heap size inherited from upstream is
64K. This is not enough to decompress the PKCS#11 early TA, as shown by
the following command:

 ledge-qemuarm64:/home/ledge# xtest pkcs11_1001
 * pkcs11_1001 PKCS11: List PKCS#11 slots and get information from
 ERR [414] LT:ckteec_invoke_init:304: TEEC open session failed ffff0000 from 3

 E/LD:  init_elf:439 sys_open_ta_bin(cb3e5ba0-adf1-11e0-998b-0002a5d5c51b)
 E/TC:? 0 ldelf_init_with_ldelf:130 ldelf failed with res: 0xffff0000
 E/TC:? 0 read_compressed:190 Decompression error (-4)

Increase the TEE core heap size to 128K to fix the issue. The setting is
added to the common file for all platforms (optee-os_%.bbappend) because
at this point there is no reason to make it platform-specific.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>